backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
pydantic-settings==2.5.2
sqlalchemy==2.0.35
pymysql==1.1.1
redis==5.0.8
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx==0.27.2
Pillow==10.4.0
python-dateutil==2.9.0.post0
openpyxl==3.1.5

croniter==3.0.3
